+++ src/compas_viewer/components/camerasetting.py
@@ -3,7 +3,7 @@ from PySide6.QtWidgets import QPushButton
 from PySide6.QtWidgets import QVBoxLayout
 
 from compas_viewer.base import Base
-from compas_viewer.components.layout import base_layout
+from compas_viewer.components.layout import SettingLayout
 
 
 class CameraSettingsDialog(QDialog, Base):
@@ -11,6 +11,11 @@ class CameraSettingsDialog(QDialog, Base):
     A dialog for displaying and updating camera settings in Qt applications.
     This dialog allows users to modify the camera's target and position and
     applies these changes dynamically.
+
+    Parameters
+    ----------
+    items : list
+        A list of dictionaries containing the settings for the camera.
 
     Attributes
     ----------
@@ -30,38 +35,19 @@ class CameraSettingsDialog(QDialog, Base):
 
     Example
     -------
-    >>> dialog = CameraSettingsDialog()
+    >>> dialog = CameraSettingsDialog(items=items)
     >>> dialog.exec()
     """
 
-    def __init__(self) -> None:
+    def __init__(self, items: list[dict]) -> None:
         super().__init__()
         self.setWindowTitle("Camera Settings")
 
         self.layout = QVBoxLayout(self)
-        self.camera = self.viewer.renderer.camera
-        items = [
-            {
-                "title": "Camera_Target",
-                "items": [
-                    {"type": "double_edit", "title": "X", "value": self.camera.target.x, "min_val": None, "max_val": None},
-                    {"type": "double_edit", "title": "Y", "value": self.camera.target.y, "min_val": None, "max_val": None},
-                    {"type": "double_edit", "title": "Z", "value": self.camera.target.z, "min_val": None, "max_val": None},
-                ],
-            },
-            {
-                "title": "Camera_Position",
-                "items": [
-                    {"type": "double_edit", "title": "X", "value": self.camera.position.x, "min_val": None, "max_val": None},
-                    {"type": "double_edit", "title": "Y", "value": self.camera.position.y, "min_val": None, "max_val": None},
-                    {"type": "double_edit", "title": "Z", "value": self.camera.position.z, "min_val": None, "max_val": None},
-                ],
-            },
-        ]
+        self.setting_layout = SettingLayout(viewer=self.viewer, items=items, type="camera_setting")
+        self.setting_layout.generate_layout()
 
-        camera_setting_layout, self.spin_boxes = base_layout(items)
-
-        self.layout.addLayout(camera_setting_layout)
+        self.layout.addLayout(self.setting_layout.layout)
 
         self.update_button = QPushButton("Update Camera", self)
         self.update_button.clicked.connect(self.update)
@@ -69,13 +55,13 @@ class CameraSettingsDialog(QDialog, Base):
 
     def update(self) -> None:
         self.viewer.renderer.camera.target.set(
-            self.spin_boxes["Camera_Target_X"].spinbox.value(),
-            self.spin_boxes["Camera_Target_Y"].spinbox.value(),
-            self.spin_boxes["Camera_Target_Z"].spinbox.value(),
+            self.setting_layout.widgets["Camera_Target_X_double_edit"].spinbox.value(),
+            self.setting_layout.widgets["Camera_Target_Y_double_edit"].spinbox.value(),
+            self.setting_layout.widgets["Camera_Target_Z_double_edit"].spinbox.value(),
         )
         self.viewer.renderer.camera.position.set(
-            self.spin_boxes["Camera_Position_X"].spinbox.value(),
-            self.spin_boxes["Camera_Position_Y"].spinbox.value(),
-            self.spin_boxes["Camera_Position_Z"].spinbox.value(),
+            self.setting_layout.widgets["Camera_Position_X_double_edit"].spinbox.value(),
+            self.setting_layout.widgets["Camera_Position_Y_double_edit"].spinbox.value(),
+            self.setting_layout.widgets["Camera_Position_Z_double_edit"].spinbox.value(),
         )
         self.accept()