+++ src/compas_viewer/components/label.py
@@ -55,6 +55,14 @@ class LabelWidget(QtWidgets.QWidget):
         self.update_minimum_size()
 
     @property
+    def default_layout(self):
+        if self._default_layout is None:
+            from compas_viewer.components.layout import DefaultLayout
+
+            self._default_layout = DefaultLayout(QtWidgets.QHBoxLayout())
+        return self._default_layout
+
+    @property
     def text(self):
         return self.label.text()