+++ src/compas_viewer/components/objectsetting.py
@@ -1,57 +1,21 @@
 from typing import TYPE_CHECKING
 
+from PySide6.QtCore import Qt
 from PySide6.QtCore import Signal
 from PySide6.QtWidgets import QDialog
 from PySide6.QtWidgets import QPushButton
+from PySide6.QtWidgets import QScrollArea
 from PySide6.QtWidgets import QVBoxLayout
 from PySide6.QtWidgets import QWidget
 
 from compas_viewer.base import Base
-from compas_viewer.components.layout import base_layout
+from compas_viewer.components.double_edit import DoubleEdit
+from compas_viewer.components.label import LabelWidget
+from compas_viewer.components.layout import SettingLayout
+from compas_viewer.components.textedit import TextEdit
 
 if TYPE_CHECKING:
     from compas_viewer import Viewer
-
-
-def object_setting_layout(viewer: "Viewer"):
-    """
-    Generates a layout for displaying and editing object information based on the selected objects in the viewer.
-
-    Parameters
-    ----------
-    viewer : Viewer
-        The viewer instance containing the scene and objects.
-
-    Returns
-    -------
-    QVBoxLayout
-        The layout for displaying object information, or None if no objects are selected.
-
-    Example
-    -------
-    >>> layout = object_setting_layout(viewer)
-    """
-    status = False
-    items = []
-    for obj in viewer.scene.objects:
-        if obj.is_selected:
-            status = True
-            new_items = [
-                {"title": "Name", "items": [{"type": "label", "text": str(obj.name)}]},
-                {"title": "Parent", "items": [{"type": "label", "text": str(obj.parent)}]},
-                {"title": "Point_Color", "items": [{"type": "color_combobox", "obj": obj, "attr": "pointcolor"}]},
-                {"title": "Line_Color", "items": [{"type": "color_combobox", "obj": obj, "attr": "linecolor"}]},
-                {"title": "Face_Color", "items": [{"type": "color_combobox", "obj": obj, "attr": "facecolor"}]},
-                {"title": "Line_Width", "items": [{"type": "double_edit", "title": "", "value": obj.linewidth, "min_val": 0.0, "max_val": 10.0}]},
-                {"title": "Point_Size", "items": [{"type": "double_edit", "title": "", "value": obj.pointsize, "min_val": 0.0, "max_val": 10.0}]},
-                {"title": "Opacity", "items": [{"type": "double_edit", "title": "", "value": obj.opacity, "min_val": 0.0, "max_val": 1.0}]},
-            ]
-            items.extend(new_items)
-
-    if not status:
-        return None
-
-    return base_layout(items)
 
 
 class ObjectSetting(QWidget):
@@ -62,17 +26,19 @@ class ObjectSetting(QWidget):
     ----------
     viewer : Viewer
         The viewer instance containing the objects.
+    items : list
+        A list of dictionaries containing the settings for the object.
 
     Attributes
     ----------
     viewer : Viewer
         The viewer instance.
+    items : list
+        A list of dictionaries containing the settings for the object.
     layout : QVBoxLayout
         The main layout for the widget.
     update_button : QPushButton
         The button to trigger the object update.
-    spin_boxes : dict
-        Dictionary to hold spin boxes for object properties.
 
     Methods
     -------
@@ -86,11 +52,23 @@ class ObjectSetting(QWidget):
 
     update_requested = Signal()
 
-    def __init__(self, viewer: "Viewer"):
+    def __init__(self, viewer: "Viewer", items: list[dict]):
         super().__init__()
         self.viewer = viewer
-        self.layout = QVBoxLayout(self)
-        self.spin_boxes = {}
+        self.items = items
+        self.setting_layout = SettingLayout(viewer=self.viewer, items=self.items, type="obj_setting")
+        # Main layout
+        self.main_layout = QVBoxLayout(self)
+
+        # Scroll area setup
+        self.scroll_area = QScrollArea(self)
+        self.scroll_area.setWidgetResizable(True)
+        self.scroll_content = QWidget()
+        self.scroll_layout = QVBoxLayout(self.scroll_content)
+        self.scroll_layout.setAlignment(Qt.AlignTop)
+        self.scroll_area.setWidget(self.scroll_content)
+
+        self.main_layout.addWidget(self.scroll_area)
 
     def clear_layout(self, layout):
         """Clear all widgets from the layout."""
@@ -106,24 +84,27 @@ class ObjectSetting(QWidget):
 
     def update(self):
         """Update the layout with the latest object settings."""
-        self.clear_layout(self.layout)
-        output = object_setting_layout(self.viewer)
+        self.clear_layout(self.scroll_layout)
+        self.setting_layout.generate_layout()
 
-        if output is not None:
-            text = "Update Object"
-            obj_setting_layout, self.spin_boxes = output
-            self.layout.addLayout(obj_setting_layout)
-            self.update_button = QPushButton(text, self)
-            self.update_button.clicked.connect(self.obj_update)
-            self.layout.addWidget(self.update_button)
+        if len(self.setting_layout.widgets) != 0:
+            self.scroll_layout.addLayout(self.setting_layout.layout)
+            for _, widget in self.setting_layout.widgets.items():
+                if isinstance(widget, DoubleEdit):
+                    widget.spinbox.valueChanged.connect(self.obj_update)
+                elif isinstance(widget, TextEdit):
+                    widget.text_edit.textChanged.connect(self.obj_update)
+        else:
+            self.scroll_layout.addWidget(LabelWidget(text="No object Selected", alignment="center"))
 
     def obj_update(self):
         """Apply the settings from spin boxes to the selected objects."""
         for obj in self.viewer.scene.objects:
             if obj.is_selected:
-                obj.linewidth = self.spin_boxes["Line_Width_"].spinbox.value()
-                obj.pointsize = self.spin_boxes["Point_Size_"].spinbox.value()
-                obj.opacity = self.spin_boxes["Opacity_"].spinbox.value()
+                obj.name = self.setting_layout.widgets["Name_text_edit"].text_edit.toPlainText()
+                obj.linewidth = self.setting_layout.widgets["Line_Width_double_edit"].spinbox.value()
+                obj.pointsize = self.setting_layout.widgets["Point_Size_double_edit"].spinbox.value()
+                obj.opacity = self.setting_layout.widgets["Opacity_double_edit"].spinbox.value()
                 obj.update()
 
 
@@ -133,12 +114,17 @@ class ObjectSettingDialog(QDialog, Base):
     This dialog allows users to modify object properties such as line width, point size, and opacity,
     and applies these changes dynamically.
 
+    Parameters
+    ----------
+    items : list
+        A list of dictionaries containing the settings for the object.
+
     Attributes
     ----------
     layout : QVBoxLayout
         The layout of the dialog.
-    spin_boxes : dict
-        Dictionary containing spin boxes for adjusting object properties.
+    items : list
+        A list of dictionaries containing the settings for the object.
     update_button : QPushButton
         Button to apply changes to the selected objects.
 
@@ -153,17 +139,16 @@ class ObjectSettingDialog(QDialog, Base):
     >>> dialog.exec()
     """
 
-    def __init__(self) -> None:
+    def __init__(self, items: list[dict]) -> None:
         super().__init__()
-
+        self.items = items
         self.setWindowTitle("Object Settings")
         self.layout = QVBoxLayout(self)
-        output = object_setting_layout(self.viewer)
+        self.setting_layout = SettingLayout(viewer=self.viewer, items=self.items, type="obj_setting")
 
-        if output is not None:
+        if self.setting_layout is not None:
             text = "Update Object"
-            obj_setting_layout, self.spin_boxes = output
-            self.layout.addLayout(obj_setting_layout)
+            self.layout.addLayout(self.setting_layout.layout)
         else:
             text = "No object selected."
 
@@ -174,9 +159,9 @@ class ObjectSettingDialog(QDialog, Base):
     def obj_update(self) -> None:
         for obj in self.viewer.scene.objects:
             if obj.is_selected:
-                obj.linewidth = self.spin_boxes["Line_Width_"].spinbox.value()
-                obj.pointsize = self.spin_boxes["Point_Size_"].spinbox.value()
-                obj.opacity = self.spin_boxes["Opacity_"].spinbox.value()
+                obj.linewidth = self.setting_layout.widgets["Line_Width_double_edit"].spinbox.value()
+                obj.pointsize = self.setting_layout.widgets["Point_Size_double_edit"].spinbox.value()
+                obj.opacity = self.setting_layout.widgets["Opacity_double_edit"].spinbox.value()
                 obj.update()
 
         self.accept()