+++ src/compas_viewer/components/sceneform.py
@@ -48,6 +48,7 @@ class Sceneform(QTreeWidget):
         self.setColumnCount(len(columns))
         self.setHeaderLabels(col["title"] for col in self.columns)
         self.setHeaderHidden(not show_headers)
+        self.setSelectionMode(QTreeWidget.SingleSelection)
         self._sceneobjects = []
 
         self.callback = callback
@@ -134,6 +135,8 @@ class Sceneform(QTreeWidget):
                 node.is_selected = node in selected_nodes
                 if self.callback and node.is_selected:
                     self.callback(self, node)
+
+            self.viewer.ui.sidebar.update()
 
         self.viewer.renderer.update()