+++ src/compas_viewer/components/slider.py
@@ -8,6 +8,8 @@ from PySide6.QtWidgets import QSlider
 from PySide6.QtWidgets import QVBoxLayout
 from PySide6.QtWidgets import QWidget
 
+from compas_viewer.components.textedit import TextEdit
+
 
 class Slider(QWidget):
     def __init__(
@@ -17,7 +19,6 @@ class Slider(QWidget):
         max_val: float = 100,
         step: Optional[float] = 1,
         action: Callable = None,
-        horizontal: Optional[bool] = True,
         starting_val: Optional[float] = None,
         tick_interval: Optional[float] = None,
     ):
@@ -72,43 +73,54 @@ class Slider(QWidget):
         super().__init__()
         self.title = title
         self.action = action
-        self._horizontal = horizontal
         self.min_val = min_val
         self.max_val = max_val
         self.step = step or 1
         self.starting_val = starting_val if starting_val is not None else self.min_val
         self._tick_interval = tick_interval if tick_interval is not None else (self._scaled_max_val - self._scaled_min_val) / 10
 
-        orientation = Qt.Horizontal if horizontal else Qt.Vertical
+        self._updating = False
 
-        self.layout = QVBoxLayout(self)
-        self._h_layout = QHBoxLayout()
-        self.slider = QSlider(orientation)
+        self._default_layout = None
+        self.layout = self.default_layout
+
+        self._text_layout = QHBoxLayout()
+        self._domain_layout = QHBoxLayout()
+        self.slider = QSlider(Qt.Horizontal)
         self.slider.setMinimum(self._scaled_min_val)
         self.slider.setMaximum(self._scaled_max_val)
         self.slider.setTickInterval(self._tick_interval)
         self.slider.setTickPosition(QSlider.TicksBelow)
         self.slider.setValue(self.starting_val)
+        # Connect the slider movement to the callback
+        self.slider.valueChanged.connect(self.on_value_changed)
 
         # Labels for displaying the range and current value
         self._min_label = QLabel(str(self.min_val), alignment=Qt.AlignLeft)
         self._max_label = QLabel(str(self.max_val), alignment=Qt.AlignRight)
         self.value_label = QLabel(f"{self.title}:")
+        self.text_edit = TextEdit(str(self.starting_val))
+        self.text_edit.text_edit.textChanged.connect(self.text_update)
 
-        # Connect the slider movement to the callback
-        self.slider.valueChanged.connect(self.on_value_changed)
-        self.layout.addWidget(self.value_label)
-        self.layout.addWidget(self.slider)
+        self._text_layout.addWidget(self.value_label)
+        self._text_layout.addWidget(self.text_edit.text_edit)
 
         # Add widgets to layout
-        if orientation == Qt.Horizontal:
-            self._h_layout.addWidget(self._min_label)
-            self._h_layout.addWidget(self._max_label)
-        else:
-            self._h_layout.addWidget(self._min_label, 0, Qt.AlignTop)
-            self._h_layout.addWidget(self._max_label, 0, Qt.AlignBottom)
+        self._domain_layout.addWidget(self._min_label)
+        self._domain_layout.addWidget(self._max_label)
 
-        self.layout.addLayout(self._h_layout)
+        self.layout.addLayout(self._text_layout)
+        self.layout.addWidget(self.slider)
+        self.layout.addLayout(self._domain_layout)
+        self.setLayout(self.layout)
+
+    @property
+    def default_layout(self):
+        if self._default_layout is None:
+            from compas_viewer.components.layout import DefaultLayout
+
+            self._default_layout = DefaultLayout(QVBoxLayout())
+        return self._default_layout
 
     @property
     def _scaled_min_val(self):
@@ -119,8 +131,24 @@ class Slider(QWidget):
         return self.max_val / self.step
 
     def on_value_changed(self, value):
-        """
-        Update the label based on the slider's current value.
-        """
-        self.value_label.setText(f"{self.title}: {round(value * self.step, 2)}")
-        self.action(self, value * self.step)
+        if self._updating:
+            return
+        self._updating = True
+        scaled_value = round(value * self.step, 2)
+        self.text_edit.text_edit.setText(str(scaled_value))
+        if self.action:
+            self.action(self, scaled_value)
+        self._updating = False
+
+    def text_update(self):
+        if self._updating:
+            return
+        self._updating = True
+        try:
+            value = float(self.text_edit.text_edit.toPlainText()) / self.step
+            self.slider.setValue(value)
+            if self.action:
+                self.action(self, value * self.step)
+        except ValueError:
+            pass  # Handle cases where the text is not a valid number
+        self._updating = False