+++ src/compas_viewer/ui/sidebar.py
@@ -5,15 +5,10 @@ from PySide6 import QtCore
 from PySide6.QtWidgets import QSplitter
 
 from compas_viewer.components import Sceneform
-from compas_viewer.components import Treeform
 from compas_viewer.components.objectsetting import ObjectSetting
 
 if TYPE_CHECKING:
     from .ui import UI
-
-
-def is_layout_empty(layout):
-    return layout.count() == 0
 
 
 class SideBarRight:
@@ -22,8 +17,8 @@ class SideBarRight:
         self.widget = QSplitter(QtCore.Qt.Orientation.Vertical)
         self.widget.setChildrenCollapsible(True)
         self.show = show
+        self.hide_widget = True
         self.items = items
-        self.sceneform = None
 
     def add_items(self) -> None:
         if not self.items:
@@ -34,25 +29,25 @@ class SideBarRight:
 
             if itemtype == "Sceneform":
                 columns = item.get("columns", None)
-                callback = item.get("callback", None)
                 if columns is None:
-                    raise ValueError("Columns not provided for Sceneform")
-                self.sceneform = Sceneform(columns, callback=callback)
+                    raise ValueError("Please setup config for Sceneform")
+                self.sceneform = Sceneform(columns=columns)
                 self.widget.addWidget(self.sceneform)
 
-            elif itemtype == "Treeform":
-                item.pop("type")
-                self.widget.addWidget(Treeform(**item))
+            elif itemtype == "ObjectSetting":
+                items = item.get("items", None)
+                if items is None:
+                    raise ValueError("Please setup config for ObjectSetting")
+                self.object_setting = ObjectSetting(viewer=self.ui.viewer, items=items)
+                self.widget.addWidget(self.object_setting)
+
+        self.show_sceneform = True
+        self.show_objectsetting = True
 
     def update(self):
         self.widget.update()
         for widget in self.widget.children():
             widget.update()
-            if isinstance(widget, ObjectSetting):
-                if is_layout_empty(widget.layout):
-                    widget.hide()
-                else:
-                    widget.show()
 
     @property
     def show(self):
@@ -60,7 +55,20 @@ class SideBarRight:
 
     @show.setter
     def show(self, value: bool):
-        if value:
-            self.widget.setVisible(True)
-        elif not value:
-            self.widget.setHidden(True)
+        self.widget.setVisible(value)
+
+    @property
+    def show_sceneform(self):
+        return self.sceneform.isVisible()
+
+    @show_sceneform.setter
+    def show_sceneform(self, value: bool):
+        self.sceneform.setVisible(value)
+
+    @property
+    def show_objectsetting(self):
+        return self.object_setting.isVisible()
+
+    @show_objectsetting.setter
+    def show_objectsetting(self, value: bool):
+        self.object_setting.setVisible(value)